requirements.txt
code128==0.2b2
csvImporter==0.1.3.2
diff-match-patch==20121119
Django==1.9.1
django-excel==0.0.10
django-import-export==1.0.0
django-sampledatahelper==0.4.1
et-xmlfile==1.0.1
jdcal==1.3
lml==0.0.1
odfpy==1.3.6
openpyxl==2.5.0
Pillow==5.0.0
pyexcel==0.5.7
pyexcel-io==0.5.6
pyexcel-webio==0.1.4
pyexcel-xls==0.5.5
pyExcelerator==0.6.4.1
pytz==2018.3
PyYAML==3.12
reportlab==3.4.0
sampledata==0.3.7
six==1.11.0
tablib==0.12.1
texttable==1.2.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
